@@ -23,6 +23,7 @@
   import { currentUrlReplaceAssetId } from '$lib/utils/navigation';
   import { TUNABLES } from '$lib/utils/tunables';
   import { thumbhash } from '$lib/actions/thumbhash';
+  import { mobileDevice } from '$lib/stores/mobile-device.svelte';
 
   interface Props {
     asset: AssetResponseDto;
@@ -71,6 +72,7 @@
     IMAGE_THUMBNAIL: { THUMBHASH_FADE_DURATION },
   } = TUNABLES;
 
+  let usingMobileDevice = $derived(mobileDevice.pointerCoarse);
   let focussableElement: HTMLElement | undefined = $state();
   let mouseOver = $state(false);
   let loaded = $state(false);
@@ -99,6 +101,7 @@
     }
     onClick?.($state.snapshot(asset));
   };
+
   const handleClick = (e: MouseEvent) => {
     if (e.ctrlKey || e.metaKey) {
       return;
@@ -109,6 +112,9 @@
   };
 
   const onMouseEnter = () => {
+    if (usingMobileDevice) {
+      return;
+    }
     mouseOver = true;
     onMouseEvent?.({ isMouseOver: true, selectedGroupIndex: groupIndex });
   };
@@ -116,6 +122,25 @@
   const onMouseLeave = () => {
     mouseOver = false;
   };
+
+  function longPress(element: HTMLElement, { onLongPress }: { onLongPress: () => void }) {
+    let timer: ReturnType<typeof setTimeout>;
+    const start = (event: TouchEvent) => {
+      timer = setTimeout(() => {
+        onLongPress();
+        event.preventDefault();
+      }, 350);
+    };
+    const end = () => clearTimeout(timer);
+    element.addEventListener('touchstart', start);
+    element.addEventListener('touchend', end);
+    return {
+      destroy: () => {
+        element.removeEventListener('touchstart', start);
+        element.removeEventListener('touchend', end);
+      },
+    };
+  }
 </script>
 
 <div
@@ -136,8 +161,12 @@
     ></canvas>
   {/if}
 
-  <!-- svelte queries for all links on afterNavigate, leading to performance problems in asset-grid which updates
-     the navigation url on scroll. Replace this with button for now. -->
+  <!-- as of iOS17, there is a preference for long press speed, which is not available for mobile web.
+      The defaults are as follows:
+      fast: 200ms
+      default: 500ms
+      slow: ??ms
+      -->
   <div
     class="group"
     style:width="{width}px"
@@ -146,6 +175,7 @@
     class:cursor-pointer={!disabled}
     onmouseenter={onMouseEnter}
     onmouseleave={onMouseLeave}
+    use:longPress={{ onLongPress: () => onSelect?.($state.snapshot(asset)) }}
     onkeydown={(evt) => {
       if (evt.key === 'Enter') {
         callClickHandlers();
@@ -161,7 +191,7 @@
     onfocus={handleFocus}
     data-testid="container-with-tabindex"
   >
-    {#if mouseOver && !disableMouseOver}
+    {#if !usingMobileDevice && mouseOver && !disableMouseOver}
       <!-- lazy show the url on mouse over-->
       <a
         class="absolute z-30 {className} top-[41px]"
@@ -208,11 +238,12 @@
       class:rounded-xl={selected}
     >
       <!-- Gradient overlay on hover -->
-      <div
-        class="absolute z-10 h-full w-full bg-gradient-to-b from-black/25 via-[transparent_25%] opacity-0 transition-opacity group-hover:opacity-100"
-        class:rounded-xl={selected}
-      ></div>
-
+      {#if !usingMobileDevice}
+        <div
+          class="absolute z-10 h-full w-full bg-gradient-to-b from-black/25 via-[transparent_25%] opacity-0 transition-opacity group-hover:opacity-100"
+          class:rounded-xl={selected}
+        ></div>
+      {/if}
       <!-- Outline on focus -->
       <div
         class="absolute size-full group-focus-visible:outline outline-4 -outline-offset-4 outline-immich-primary"
