@@ -1,13 +1,8 @@
 <script lang="ts">
   import ImmichLogo from '$lib/components/shared-components/immich-logo.svelte';
-
-  interface Props {
-    width: number;
-  }
-
-  let { width }: Props = $props();
+  import { mobileDevice } from '$lib/stores/mobile-device.svelte';
 </script>
 
 <a data-sveltekit-preload-data="hover" class="ml-4" href="/">
-  <ImmichLogo class="h-[24px] w-[24px] max-w-none md:w-auto md:h-10 md:max-w-full" noText={width < 768} />
+  <ImmichLogo class="h-[50px] max-w-none md:w-auto md:max-w-full" noText={mobileDevice.maxMd} />
 </a>
