@@ -164,8 +164,21 @@
   };
 
   function getSearchType(): 'smart' | 'metadata' | 'description' {
-    const t = localStorage.getItem('searchQueryType');
-    return t === 'smart' || t === 'description' ? t : 'metadata';
+    const searchType = localStorage.getItem('searchQueryType');
+    switch (searchType) {
+      case 'smart': {
+        return 'smart';
+      }
+      case 'metadata': {
+        return 'metadata';
+      }
+      case 'description': {
+        return 'description';
+      }
+      default: {
+        return 'smart';
+      }
+    }
   }
 
   function getSearchTypeText(): string {
@@ -209,7 +222,7 @@
         type="text"
         name="q"
         id="main-search-bar"
-        class="w-full transition-all border-2 px-14 py-4 text-immich-fg/75 dark:text-immich-dark-fg
+        class="w-full transition-all border-2 px-14 py-4 max-md:py-2 text-immich-fg/75 dark:text-immich-dark-fg
         {grayTheme ? 'dark:bg-immich-dark-gray' : 'dark:bg-immich-dark-bg'}
         {showSuggestions && isSearchSuggestions ? 'rounded-t-3xl' : 'rounded-3xl bg-gray-200'}
         {$isSearchEnabled && !showFilter ? 'border-gray-200 dark:border-gray-700 bg-white' : 'border-transparent'}"
