@@ -71,7 +71,9 @@
 
 <div class="mb-4 w-full">
   <div class="flex place-items-center gap-1">
-    <label class="font-medium text-immich-primary dark:text-immich-dark-primary text-sm" for={label}>{label}</label>
+    <label class="font-medium text-immich-primary dark:text-immich-dark-primary text-sm min-h-6" for={label}
+      >{label}</label
+    >
     {#if required}
       <div class="text-red-400">*</div>
     {/if}
@@ -109,7 +111,7 @@
           max={max.toString()}
           {step}
           {required}
-          {value}
+          bind:value
           onchange={handleChange}
           {disabled}
           {title}
@@ -129,7 +131,7 @@
         max={max.toString()}
         {step}
         {required}
-        {value}
+        bind:value
         onchange={handleChange}
         {disabled}
         {title}
