@@ -1,9 +1,13 @@
 import { MediaQuery } from 'svelte/reactivity';
 
-const hoverNone = new MediaQuery('hover: none');
+const pointerCoarse = new MediaQuery('pointer:coarse');
+const maxMd = new MediaQuery('max-width: 767px');
 
 export const mobileDevice = {
-  get hoverNone() {
-    return hoverNone.current;
+  get pointerCoarse() {
+    return pointerCoarse.current;
+  },
+  get maxMd() {
+    return maxMd.current;
   },
 };
