@@ -394,7 +394,7 @@
     }
   });
 
-  let album = $state(data.album);
+  let album = $derived(data.album);
   let albumId = $derived(album.id);
 
   $effect(() => {
@@ -404,6 +404,7 @@
   });
 
   let assetStore = new AssetStore();
+
   $effect(() => {
     if (viewMode === AlbumPageViewMode.VIEW) {
       void assetStore.updateOptions({ albumId, order: albumOrder });
@@ -603,7 +604,9 @@
       {/if}
     {/if}
 
-    <main class="relative h-screen overflow-hidden bg-immich-bg px-6 pt-[var(--navbar-height)] dark:bg-immich-dark-bg">
+    <main
+      class="relative h-screen overflow-hidden bg-immich-bg px-6 max-md:pt-[var(--navbar-height-md)] pt-[var(--navbar-height)] dark:bg-immich-dark-bg"
+    >
       <AssetGrid
         enableRouting={viewMode === AlbumPageViewMode.SELECT_ASSETS ? false : true}
         {album}
